# src/main/resources/static/api-docs.yaml
openapi: 3.0.3
info:
  title: Cinema Ticketing API
  version: 2.0.0
  description: API для системы продажи билетов в кинотеатр.
servers:
  - url: http://91.142.94.183:8080
  - url: http://localhost:8080
paths:
  /auth/register:
    post:
      summary: Регистрация пользователя
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Ошибка валидации
        '409':
          description: Пользователь уже существует
  /auth/login:
    post:
      summary: Логин пользователя
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Неверные данные
...
  /films:
    get:
      summary: Получить список фильмов
      tags: [Films]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Список фильмов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilmListResponse'
    post:
      summary: Создать фильм (admin)
      tags: [Films]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilmCreate'
      responses:
        '201':
          description: Фильм создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Film'
  /films/{id}:
    get:
      summary: Получить фильм по ID
      tags: [Films]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Фильм
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Film'
        '404':
          description: Не найден

    put:
      summary: Обновить фильм (admin)
      tags: [Films]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilmUpdate'
      responses:
        '200':
          description: Фильм обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Film'
        '404':
          description: Фильм не найден

    delete:
      summary: Удалить фильм (admin)
      tags: [Films]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Фильм удален
        '404':
          description: Фильм не найден
...
  /halls:
    get:
      summary: Получить список залов
      tags: [Halls]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Список залов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HallListResponse'
    post:
      summary: Создать зал с планом (admin)
      description: Создает новый зал с обязательным планом расположения мест
      tags: [Halls]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HallCreate'
      responses:
        '201':
          description: Зал создан с планом расположения мест
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hall'
        '400':
          description: Ошибка валидации (отсутствуют места или неверные данные)
  /halls/{id}:
    get:
      summary: Получить зал по ID
      tags: [Halls]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Зал
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hall'
        '404':
          description: Зал не найден

    put:
      summary: Обновить зал (admin)
      tags: [Halls]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HallUpdate'
      responses:
        '200':
          description: Зал обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hall'
        '404':
          description: Зал не найден

    delete:
      summary: Удалить зал (admin)
      tags: [Halls]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Зал удален
        '404':
          description: Зал не найден
  /halls/{id}/plan:
    get:
      summary: Получить план зала
      tags: [Halls]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: План зала
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HallPlan'
        '404':
          description: Зал не найден

    put:
      summary: Обновить план зала (admin)
      tags: [Halls]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HallPlanUpdate'
      responses:
        '200':
          description: План зала обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HallPlan'
        '404':
          description: Зал не найден
...
  /purchases:
    get:
      summary: Получить историю покупок пользователя
      tags: [Purchases]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: История покупок
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseListResponse'

    post:
      summary: Создать покупку
      tags: [Purchases]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseCreate'
      responses:
        '201':
          description: Покупка создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Purchase'
        '400':
          description: Ошибка валидации или билеты недоступны
  /purchases/{id}:
    get:
      summary: Получить покупку по ID
      tags: [Purchases]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Покупка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Purchase'
        '404':
          description: Покупка не найдена
  /purchases/{id}/cancel:
    post:
      summary: Отменить покупку
      tags: [Purchases]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Покупка отменена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Purchase'
        '400':
          description: Невозможно отменить покупку
...
  /payments/process:
    post:
      summary: Обработать платеж (заглушка)
      tags: [Payments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '200':
          description: Платеж обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Ошибка обработки платежа
  /payments/{id}/status:
    get:
      summary: Получить статус платежа
      tags: [Payments]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Статус платежа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentStatus'
        '404':
          description: Платеж не найден
  /media/upload:
    post:
      summary: Загрузить медиафайл (admin)
      description: Загружает медиафайл в базу данных как binary данные (PostgreSQL BYTEA)
      tags: [Media]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Медиафайл для загрузки
                mediaType:
                  type: string
                  enum: [IMAGE, AUDIO, VIDEO]
                  description: Тип медиафайла
              required: [file, mediaType]
      responses:
        '201':
          description: Медиафайл загружен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'
        '400':
          description: Ошибка валидации
        '413':
          description: Файл слишком большой
  /media/{id}:
    get:
      summary: Получить медиафайл
      tags: [Media]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Медиафайл
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Медиафайл не найден

    delete:
      summary: Удалить медиафайл (admin)
      tags: [Media]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Медиафайл удален
        '404':
          description: Медиафайл не найден
components:
  schemas:
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        firstName:
          type: string
        lastName:
          type: string
        age:
          type: integer
          minimum: 0
        gender:
          type: string
          enum: [MALE, FEMALE]
          description: Пол пользователя
      required: [email, password, firstName, lastName, gender]
...
    FilmListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Film'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        pages: { type: integer }
    FilmUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        durationMinutes: { type: integer }
        ageRating:
          type: string
          enum: ["0+", "6+", "12+", "16+", "18+"]
        posterId: { type: string }
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
    Film:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        durationMinutes: { type: integer }
        ageRating:
          type: string
          enum: ["0+", "6+", "12+", "16+", "18+"]
        poster:
          $ref: '#/components/schemas/Media'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Media:
      type: object
      properties:
        id: { type: string }
        filename: { type: string }
        contentType: { type: string }
        mediaType:
          type: string
          enum: [IMAGE, AUDIO, VIDEO]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    FilmCreate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        durationMinutes: { type: integer }
        ageRating:
          type: string
          enum: ["0+", "6+", "12+", "16+", "18+"]
        posterId: { type: string }
      required: [title, description, durationMinutes, ageRating]
...
    Purchase:
      type: object
      properties:
        id: { type: string }
        clientId: { type: string }
        filmId: { type: string }
        ticketIds:
          type: array
          items: { type: string }
        totalCents: { type: integer }
        status:
          type: string
          enum: [PENDING, PAID, FAILED, CANCELLED]
        createdAt: { type: string, format: date-time }
...
    Hall:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        number: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SeatCategory:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        priceCents: { type: integer }
    Seat:
      type: object
      properties:
        id: { type: string }
        row: { type: integer }
        number: { type: integer }
        categoryId: { type: string }
        status:
          type: string
          enum: [AVAILABLE, RESERVED, SOLD, CANCELLED]
    HallPlan:
      type: object
      properties:
        hallId: { type: string }
        rows: { type: integer }
        seats:
          type: array
          items:
            $ref: '#/components/schemas/Seat'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/SeatCategory'
...
    HallListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Hall'
        pagination:
          $ref: '#/components/schemas/Pagination'
    HallCreate:
      type: object
      description: Создание зала с обязательным планом расположения мест
      properties:
        name: 
          type: string
          description: Название зала
        number: 
          type: integer
          description: Номер зала
          minimum: 1
        rows: 
          type: integer
          description: Количество рядов в зале
          minimum: 1
        seats:
          type: array
          description: Список мест в зале (обязательный, не может быть пустым)
          minItems: 1
          items:
            $ref: '#/components/schemas/SeatCreate'
      required: [name, number, rows, seats]
    HallUpdate:
      type: object
      properties:
        name: { type: string }
        number: { type: integer }
    HallPlanUpdate:
      type: object
      properties:
        rows: { type: integer }
        seats:
          type: array
          items:
            $ref: '#/components/schemas/SeatCreate'
      required: [rows, seats]
    SeatCreate:
      type: object
      properties:
        row: { type: integer }
        number: { type: integer }
        categoryId: { type: string }
      required: [row, number, categoryId]
...
    PurchaseListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Purchase'
        pagination:
          $ref: '#/components/schemas/Pagination'
    PurchaseCreate:
      type: object
      properties:
        ticketIds:
          type: array
          items: { type: string }
      required: [ticketIds]
...
    PaymentRequest:
      type: object
      properties:
        purchaseId: { type: string }
        cardNumber: { type: string }
        expiryDate: { type: string }
        cvv: { type: string }
        cardHolderName: { type: string }
      required: [purchaseId, cardNumber, expiryDate, cvv, cardHolderName]
    PaymentResponse:
      type: object
      properties:
        paymentId: { type: string }
        status:
          type: string
          enum: [SUCCESS, FAILED, PENDING]
        message: { type: string }
    PaymentStatus:
      type: object
      properties:
        paymentId: { type: string }
        status:
          type: string
          enum: [SUCCESS, FAILED, PENDING]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
  parameters:
    page:
      name: page
      in: query
      description: Номер страницы
      schema:
        type: integer
        default: 0
        minimum: 0
    limit:
      name: size
      in: query
      description: Количество элементов на странице
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT